requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
exceptiongroup==1.3.0
fastapi==0.118.0
h11==0.16.0
idna==3.10
kombu==5.5.4
mailtrap==2.2.0
packaging==25.0
prompt_toolkit==3.0.52
pydantic==2.12.0
pydantic_core==2.41.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.5
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
vine==5.1.0
wcwidth==0.2.14
